+++ conftest.py
@@ -4,6 +4,7 @@
 # 2025-04-23 - JS
 
 import warnings
+import pytest
 
 # Filter out warnings from dependencies
 warnings.filterwarnings("ignore", message="torchaudio._backend.*has been deprecated")
@@ -13,3 +14,11 @@ warnings.filterwarnings("ignore", category=DeprecationWarning, module="matplotli
 warnings.filterwarnings("ignore", category=DeprecationWarning, module="pydub.utils")
 warnings.filterwarnings("ignore", message="The get_cmap function was deprecated in Matplotlib 3.7")
 warnings.filterwarnings("ignore", message="`torchaudio.backend.common.AudioMetaData` has been moved to `torchaudio.AudioMetaData`")
+
+# Configure pytest to filter warnings
+@pytest.hookimpl(hookwrapper=True)
+def pytest_runtest_protocol(item, nextitem):
+    # This runs before each test
+    warnings.filterwarnings("ignore", category=DeprecationWarning)
+    warnings.filterwarnings("ignore", category=UserWarning)
+    yield  # This allows the test to run